Converted number to 'long double' instead of 'double'

+++ src/value.h
@@ -18,7 +18,7 @@ typedef struct {
 typedef struct {
     ValueType type;
     union {
-        double num;
+        long double num;
         String str;
         bool   bol;
     } as;
